+++ tests/io/test_factory.py
@@ -3,33 +3,27 @@ import tempfile
 import unittest
 import os
 
-import cwepr.exceptions
+import aspecd.io
+
 import cwepr.dataset
+import cwepr.exceptions
+import cwepr.io
 
 ROOTPATH = os.path.split(os.path.abspath(__file__))[0]
 
 
 class TestDatasetImporterFactory(unittest.TestCase):
-    def test_error_message_no_matching_file_pair(self):
-        source = os.path.join(ROOTPATH, 'testdata',
-                              'test-magnettech.xml').replace('xml', 'test')
-        importer = cwepr.dataset.DatasetFactory()
-        with self.assertRaises(cwepr.exceptions.UnsupportedDataFormatError) \
-                as error:
-            importer.get_dataset(source=source)
-        self.assertIn('No file format was found', error.exception.message)
+    def setUp(self):
+        self.factory = cwepr.io.DatasetImporterFactory()
+
+    def test_instantiate_class(self):
+        pass
 
     def test_goniometer_importer_gets_correct_files(self):
         source = os.path.join(ROOTPATH, 'testdata', 'magnettech-goniometer/')
         importer = cwepr.dataset.DatasetFactory().importer_factory
         importer.get_importer(source=source)
         self.assertEqual('GoniometerSweep', importer.data_format)
-
-    def test_goniometer_importer_does_not_recognise_empty_directory(self):
-        with tempfile.TemporaryDirectory() as empty_folder:
-            importer = cwepr.dataset.DatasetFactory().importer_factory
-            with self.assertRaises(cwepr.exceptions.UnsupportedDataFormatError):
-                importer.get_importer(source=empty_folder)
 
     def test_goniometer_importer_does_not_import_inconsistent_data(self):
         source = os.path.join(ROOTPATH, 'testdata', 'magnettech-goniometer/')
@@ -41,8 +35,8 @@ class TestDatasetImporterFactory(unittest.TestCase):
             target_name = os.path.join(new_source, 'fake_name')
             os.rename(source_name, target_name)
             importer = cwepr.dataset.DatasetFactory().importer_factory
-            with self.assertRaises(cwepr.exceptions.UnsupportedDataFormatError):
-                importer.get_importer(source=new_source)
+            importer.get_importer(source=new_source)
+            self.assertNotEqual('MagnettechXML', importer.data_format)
 
     def test_factory_cuts_filename(self):
         source = os.path.join(ROOTPATH, 'testdata', 'test-magnettech.xml')
@@ -51,3 +45,8 @@ class TestDatasetImporterFactory(unittest.TestCase):
         importer_factory.get_importer(source=source)
         root_source, _ = os.path.splitext(source)
         self.assertEqual(root_source, importer_factory.source)
+
+    def test_with_adf_extension_returns_adf_importer(self):
+        source = 'test.adf'
+        importer = self.factory.get_importer(source=source)
+        self.assertIsInstance(importer, aspecd.io.AdfImporter)